Fix includeField parsing
