+++ src/Microsoft.Health.Dicom.Api/Features/ModelBinders/CsvModelBinder.T.cs
@@ -4,29 +4,51 @@
 // -------------------------------------------------------------------------------------------------
 
 using System;
+using System.Threading.Tasks;
+using EnsureThat;
+using Microsoft.AspNetCore.Mvc.ModelBinding;
+using Microsoft.Extensions.Primitives;
 
 namespace Microsoft.Health.Dicom.Api.Features.ModelBinders
 {
-    internal abstract class CsvModelBinder<T> : CsvModelBinder
+    internal abstract class CsvModelBinder<T> : IModelBinder
     {
-        protected override object DefaultValue => Array.Empty<T>();
-
-        protected sealed override bool TryParse(string[] values, out object result)
+        public Task BindModelAsync(ModelBindingContext bindingContext)
         {
-            T[] parsed = new T[values.Length];
-            for (int i = 0; i < values.Length; i++)
+            EnsureArg.IsNotNull(bindingContext, nameof(bindingContext));
+            StringValues values = bindingContext.ValueProvider.GetValue(bindingContext.ModelName).Values;
+
+            if (values.Count == 0)
             {
-                if (!TryParse(values[i], out T parsedValue))
+                bindingContext.Result = ModelBindingResult.Success(Array.Empty<T>());
+            }
+            else if (values.Count > 1)
+            {
+                bindingContext.ModelState.TryAddModelError(bindingContext.ModelName, DicomApiResource.DuplicateValues);
+            }
+            else if (string.IsNullOrEmpty(values[0]))
+            {
+                bindingContext.Result = ModelBindingResult.Success(Array.Empty<T>());
+            }
+            else
+            {
+                string[] split = values[0].Split(',', StringSplitOptions.TrimEntries);
+                T[] parsed = new T[split.Length];
+                for (int i = 0; i < split.Length; i++)
                 {
-                    result = default;
-                    return false;
+                    if (!TryParse(split[i], out T parsedValue))
+                    {
+                        bindingContext.ModelState.TryAddModelError(bindingContext.ModelName, string.Format(DicomApiResource.InvalidParse, split[0], typeof(T).Name));
+                        return Task.CompletedTask;
+                    }
+
+                    parsed[i] = parsedValue;
                 }
 
-                parsed[i] = parsedValue;
+                bindingContext.Result = ModelBindingResult.Success(parsed);
             }
 
-            result = parsed;
-            return true;
+            return Task.CompletedTask;
         }
 
         protected abstract bool TryParse(string value, out T result);
